Added connectivity to themoviedb https://www.themoviedb.org/settings/api

+++ microservices/catalogserv/src/main/java/org/picokernel/catalog/resources/models/CatalogItem.java
@@ -1,13 +1,14 @@
 package org.picokernel.catalog.resources.models;
 
+import java.math.BigDecimal;
 import java.util.Objects;
 
 public class CatalogItem {
     private String name;
     private String description;
-    private int rating;
+    private BigDecimal rating;
 
-    public CatalogItem(String name, String description, int rating) {
+    public CatalogItem(String name, String description, BigDecimal rating) {
         this.name = name;
         this.description = description;
         this.rating = rating;
@@ -29,11 +30,11 @@ public class CatalogItem {
         this.description = description;
     }
 
-    public int getRating() {
+    public BigDecimal getRating() {
         return rating;
     }
 
-    public void setRating(int rating) {
+    public void setRating(BigDecimal rating) {
         this.rating = rating;
     }
